results/beyond_wpt/webspec_tt_frames/verifier/launcher.html

<!DOCTYPE html>
<meta charset=utf-8>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/resources/channel.sub.js"></script>
<script src=/websockets/constants.sub.js></script>
<script src=/portals/resources/stash-utils.sub.js></script>
<script src="/common/utils.js"></script>

<script>

const delay = (time) =>
      new Promise((resolve) => setTimeout(resolve, time))

setup({explicit_timeout: true});

const timeout_fn = async(time, label) => {
  await delay(time) 
  assert_true(false, label)
}


function run_test(timeout, label, ...fns) {
  return Promise.race(
    [ Promise.all(fns.map((x) => x())), 
      timeout_fn(timeout, label)
    ]);
}


promise_test(async t => {
  // here maybe the setup, where we install the SW (and start off the event chain by writing to the stash??)
   

  
  
  window.open('https://web-platform.test:1024/verifier/server.py?res=216');
  StashUtils.putValue('b3dad051-8c09-48b1-8a63-310b223a5f41', 'irrelevant');
  
  StashUtils.putValue('bce991e8-7650-4d16-8ea4-df5da56520f2', 'dummy');
  
   

  await run_test(10000, "Timed out", 
    
    () => StashUtils.takeValue("a49d9788-6e3b-4df9-8b4c-d9aca1f11e85").then(t.step_func(value => {
        assert_equals(value, "obghtagajq", "test 4");
    })),
    
    () => StashUtils.takeValue("b3dad051-8c09-48b1-8a63-310b223a5f41").then(t.step_func(value => {
        assert_equals(value, "irrelevant", "test 3");
    })),
    
    () => StashUtils.takeValue("21dddc12-f95d-42db-b414-3b5980bf6d0c").then(t.step_func(value => {
        assert_equals(value, "GET.https.www.not-web-platform.test.1025.204", "test 2");
    })),
    
    () => StashUtils.takeValue("d36f5855-e735-4283-8891-6cbf63156261").then(t.step_func(value => {
        assert_equals(value, "GET.https.web-platform.test.1024.219", "test 1");
    })),
    
    () => StashUtils.takeValue("29f18179-b604-4b72-8eff-ef9c9c9f1e8a").then(t.step_func(value => {
        assert_equals(value, "GET.https.web-platform.test.1024.216", "test 0");
    })),
     
    async () => { return }
  );
}, "TTInvariant.trace");

</script>
